+++ include/tatami_stats/grouped_variances.hpp
@@ -322,7 +322,7 @@ void apply(bool row, const tatami::Matrix<Value_, Index_>* p, const Group_* grou
             tatami::Options opt;
             opt.sparse_ordered_index = false; 
 
-            tatami::parallelize([&](size_t thread, Index_ start, Index_ len) -> void {
+            tatami::parallelize([&](int thread, Index_ start, Index_ len) -> void {
                 std::vector<variances::RunningSparse<Output_, Value_, Index_> > runners;
                 runners.reserve(num_groups);
                 std::vector<LocalOutputBuffer<Output_> > local_var_output;
@@ -382,7 +382,7 @@ void apply(bool row, const tatami::Matrix<Value_, Index_>* p, const Group_* grou
             }, dim, sopt.num_threads);
 
         } else {
-            tatami::parallelize([&](size_t thread, Index_ start, Index_ len) -> void {
+            tatami::parallelize([&](int thread, Index_ start, Index_ len) -> void {
                 std::vector<variances::RunningDense<Output_, Value_, Index_> > runners;
                 runners.reserve(num_groups);
                 std::vector<LocalOutputBuffer<Output_> > local_var_output;
