+++ include/tatami_stats/ranges.hpp
@@ -352,7 +352,7 @@ void apply(bool row, const tatami::Matrix<Value_, Index_>* p, Output_* min_out, 
 
         if (direct) {
             opt.sparse_extract_index = false;
-            tatami::parallelize([&](size_t, Index_ s, Index_ l) {
+            tatami::parallelize([&](int, Index_ s, Index_ l) {
                 auto ext = tatami::consecutive_extractor<true>(p, row, s, l, opt);
                 std::vector<Value_> vbuffer(otherdim);
                 for (Index_ x = 0; x < l; ++x) {
@@ -367,7 +367,7 @@ void apply(bool row, const tatami::Matrix<Value_, Index_>* p, Output_* min_out, 
             }, dim, ropt.num_threads);
 
         } else {
-            tatami::parallelize([&](size_t thread, Index_ s, Index_ l) {
+            tatami::parallelize([&](int thread, Index_ s, Index_ l) {
                 auto ext = tatami::consecutive_extractor<true>(p, !row, static_cast<Index_>(0), otherdim, s, l, opt);
                 std::vector<Value_> vbuffer(l);
                 std::vector<Index_> ibuffer(l);
@@ -400,7 +400,7 @@ void apply(bool row, const tatami::Matrix<Value_, Index_>* p, Output_* min_out, 
 
     } else {
         if (direct) {
-            tatami::parallelize([&](size_t, Index_ s, Index_ l) {
+            tatami::parallelize([&](int, Index_ s, Index_ l) {
                 auto ext = tatami::consecutive_extractor<false>(p, row, s, l);
                 std::vector<Value_> buffer(otherdim);
                 for (Index_ x = 0; x < l; ++x) {
@@ -415,7 +415,7 @@ void apply(bool row, const tatami::Matrix<Value_, Index_>* p, Output_* min_out, 
             }, dim, ropt.num_threads);
 
         } else {
-            tatami::parallelize([&](size_t thread, Index_ s, Index_ l) {
+            tatami::parallelize([&](int thread, Index_ s, Index_ l) {
                 auto ext = tatami::consecutive_extractor<false>(p, !row, static_cast<Index_>(0), otherdim, s, l);
                 std::vector<Value_> buffer(l);
